> Rename class Folders to LocalFolderTasks > Typos in release .json downloader > Small typos in logging
